% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birchmash_raw.R
\docType{data}
\name{birchmash_raw}
\alias{birchmash_raw}
\title{birchmash_raw}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 168 rows and 153 columns.
\itemize {
    \item DISK: "double" variable with 111 unique values in the range [13, 442], 58 NA values.
    \item TRAKCORE: "double" variable with 59 unique values in the range [1.1, 80.1], 110 NA values.
    \item RCNT: "double" variable with 39 unique values in the range [4, 125], 116 NA values.
    \item RW1998: "double" variable with 53 unique values in the range [0.013, 15.257], 116 NA values.
    \item RW1997: "double" variable with 53 unique values in the range [0.051, 23.401], 116 NA values.
    \item RW1996: "double" variable with 51 unique values in the range [0.027, 10.153], 116 NA values.
    \item RW1995: "double" variable with 50 unique values in the range [0.062, 17.613], 116 NA values.
    \item RW1994: "double" variable with 52 unique values in the range [0.027, 19.779], 117 NA values.
    \item RW1993: "double" variable with 51 unique values in the range [0.089, 19.409], 117 NA values.
    \item RW1992: "double" variable with 52 unique values in the range [0.05, 10.737], 117 NA values.
    \item RW1991: "double" variable with 52 unique values in the range [0.083, 10.115], 117 NA values.
    \item RW1990: "double" variable with 51 unique values in the range [0, 22.603], 118 NA values.
    \item RW1989: "double" variable with 51 unique values in the range [0.051, 24.59], 118 NA values.
    \item RW1988: "double" variable with 51 unique values in the range [0.066, 23.701], 118 NA values.
    \item RW1987: "double" variable with 51 unique values in the range [0.051, 20.252], 118 NA values.
    \item RW1986: "double" variable with 51 unique values in the range [0.101, 19.402], 118 NA values.
    \item RW1985: "double" variable with 50 unique values in the range [0.102, 9.918], 118 NA values.
    \item RW1984: "double" variable with 51 unique values in the range [0.038, 11.894], 118 NA values.
    \item RW1983: "double" variable with 50 unique values in the range [0.088, 14.141], 119 NA values.
    \item RW1982: "double" variable with 49 unique values in the range [0.076, 25.069], 119 NA values.
    \item RW1981: "double" variable with 48 unique values in the range [0.076, 28.156], 120 NA values.
    \item RW1980: "double" variable with 49 unique values in the range [0.061, 60.65], 120 NA values.
    \item RW1979: "double" variable with 49 unique values in the range [0.039, 41.601], 120 NA values.
    \item RW1978: "double" variable with 49 unique values in the range [0.038, 79.456], 120 NA values.
    \item RW1977: "double" variable with 48 unique values in the range [0.039, 31.278], 120 NA values.
    \item RW1976: "double" variable with 47 unique values in the range [0.051, 45.171], 121 NA values.
    \item RW1975: "double" variable with 48 unique values in the range [0.065, 16.678], 121 NA values.
    \item RW1974: "double" variable with 48 unique values in the range [0, 22.382], 121 NA values.
    \item RW1973: "double" variable with 48 unique values in the range [0.074, 12.698], 121 NA values.
    \item RW1972: "double" variable with 47 unique values in the range [0, 15.068], 121 NA values.
    \item RW1971: "double" variable with 48 unique values in the range [0.049, 11.633], 121 NA values.
    \item RW1970: "double" variable with 47 unique values in the range [0.061, 26.752], 121 NA values.
    \item RW1969: "double" variable with 48 unique values in the range [0.076, 11.353], 121 NA values.
    \item RW1968: "double" variable with 46 unique values in the range [0.171, 24.71], 122 NA values.
    \item RW1967: "double" variable with 46 unique values in the range [0.127, 30.429], 123 NA values.
    \item RW1966: "double" variable with 46 unique values in the range [0.069, 44.732], 123 NA values.
    \item RW1965: "double" variable with 44 unique values in the range [0.088, 15.467], 124 NA values.
    \item RW1964: "double" variable with 44 unique values in the range [0.081, 9.721], 125 NA values.
    \item RW1963: "double" variable with 42 unique values in the range [0.188, 14.129], 125 NA values.
    \item RW1962: "double" variable with 43 unique values in the range [0.069, 19.056], 126 NA values.
    \item RW1961: "double" variable with 40 unique values in the range [0.089, 21.195], 129 NA values.
    \item RW1960: "double" variable with 39 unique values in the range [0.112, 22.784], 130 NA values.
    \item RW1959: "double" variable with 38 unique values in the range [0.107, 30.831], 131 NA values.
    \item RW1958: "double" variable with 37 unique values in the range [0.03, 51.456], 132 NA values.
    \item RW1957: "double" variable with 37 unique values in the range [0.103, 45.382], 132 NA values.
    \item RW1956: "double" variable with 34 unique values in the range [0.064, 19.883], 135 NA values.
    \item RW1955: "double" variable with 31 unique values in the range [0.06, 20.119], 137 NA values.
    \item RW1954: "double" variable with 29 unique values in the range [0.059, 3.295], 140 NA values.
    \item RW1953: "double" variable with 26 unique values in the range [0.063, 2.108], 143 NA values.
    \item RW1952: "double" variable with 22 unique values in the range [0.091, 3.16], 147 NA values.
    \item RW1951: "double" variable with 22 unique values in the range [0.128, 2.216], 147 NA values.
    \item RW1950: "double" variable with 20 unique values in the range [0.139, 1.766], 149 NA values.
    \item RW1949: "double" variable with 19 unique values in the range [0.118, 1.759], 150 NA values.
    \item RW1948: "double" variable with 15 unique values in the range [0.153, 1.429], 153 NA values.
    \item RW1947: "double" variable with 13 unique values in the range [0.023, 1.848], 156 NA values.
    \item RW1946: "double" variable with 12 unique values in the range [0.152, 2.347], 157 NA values.
    \item RW1945: "double" variable with 12 unique values in the range [0.084, 2.118], 157 NA values.
    \item RW1944: "double" variable with 10 unique values in the range [0.072, 1.169], 159 NA values.
    \item RW1943: "double" variable with 9 unique values in the range [0.088, 1.724], 160 NA values.
    \item RW1942: "double" variable with 9 unique values in the range [0.132, 1.333], 160 NA values.
    \item RW1941: "double" variable with 9 unique values in the range [0.102, 1.684], 160 NA values.
    \item RW1940: "double" variable with 8 unique values in the range [0.057, 1.643], 161 NA values.
    \item RW1939: "double" variable with 8 unique values in the range [0.058, 1.516], 161 NA values.
    \item RW1938: "double" variable with 8 unique values in the range [0.137, 1.67], 161 NA values.
    \item RW1937: "double" variable with 8 unique values in the range [0.069, 1.538], 161 NA values.
    \item RW1936: "double" variable with 8 unique values in the range [0.063, 1.944], 161 NA values.
    \item RW1935: "double" variable with 8 unique values in the range [0.038, 2.165], 161 NA values.
    \item RW1934: "double" variable with 8 unique values in the range [0.038, 2.685], 161 NA values.
    \item RW1933: "double" variable with 8 unique values in the range [0.05, 3.04], 161 NA values.
    \item RW1932: "double" variable with 8 unique values in the range [0.063, 2.258], 161 NA values.
    \item RW1931: "double" variable with 8 unique values in the range [0.056, 1.903], 161 NA values.
    \item RW1930: "double" variable with 8 unique values in the range [0.081, 1.567], 161 NA values.
    \item RW1929: "double" variable with 8 unique values in the range [0.056, 1.494], 161 NA values.
    \item RW1928: "double" variable with 8 unique values in the range [0.057, 1.779], 161 NA values.
    \item RW1927: "double" variable with 8 unique values in the range [0.063, 1.628], 161 NA values.
    \item RW1926: "double" variable with 8 unique values in the range [0.05, 0.897], 161 NA values.
    \item RW1925: "double" variable with 8 unique values in the range [0.038, 1.513], 161 NA values.
    \item RW1924: "double" variable with 8 unique values in the range [0.056, 1.747], 161 NA values.
    \item RW1923: "double" variable with 8 unique values in the range [0.05, 1.612], 161 NA values.
    \item RW1922: "double" variable with 8 unique values in the range [0.019, 1.38], 161 NA values.
    \item RW1921: "double" variable with 8 unique values in the range [0.157, 1.986], 161 NA values.
    \item RW1920: "double" variable with 8 unique values in the range [0.073, 2.573], 161 NA values.
    \item RW1919: "double" variable with 7 unique values in the range [0.139, 2.556], 162 NA values.
    \item RW1918: "double" variable with 7 unique values in the range [0.144, 2.506], 162 NA values.
    \item RW1917: "double" variable with 7 unique values in the range [0.062, 1.924], 162 NA values.
    \item RW1916: "double" variable with 7 unique values in the range [0.34, 2.515], 162 NA values.
    \item RW1915: "double" variable with 7 unique values in the range [0.214, 2.234], 162 NA values.
    \item RW1914: "double" variable with 7 unique values in the range [0.249, 1.897], 162 NA values.
    \item RW1913: "double" variable with 7 unique values in the range [0.037, 1.562], 162 NA values.
    \item RW1912: "double" variable with 7 unique values in the range [0.176, 1.726], 162 NA values.
    \item RW1911: "double" variable with 6 unique values in the range [0.044, 1.272], 163 NA values.
    \item RW1910: "double" variable with 6 unique values in the range [0.15, 1.232], 163 NA values.
    \item RW1909: "double" variable with 6 unique values in the range [0.313, 1.906], 163 NA values.
    \item RW1908: "double" variable with 6 unique values in the range [0.407, 1.953], 163 NA values.
    \item RW1907: "double" variable with 6 unique values in the range [0.42, 1.519], 163 NA values.
    \item RW1906: "double" variable with 6 unique values in the range [0.218, 1.447], 163 NA values.
    \item RW1905: "double" variable with 6 unique values in the range [0.158, 1.336], 163 NA values.
    \item RW1904: "double" variable with 6 unique values in the range [0.193, 0.806], 163 NA values.
    \item RW1903: "double" variable with 6 unique values in the range [0.182, 0.742], 163 NA values.
    \item RW1902: "double" variable with 6 unique values in the range [0.207, 0.742], 163 NA values.
    \item RW1901: "double" variable with 6 unique values in the range [0.286, 1.312], 163 NA values.
    \item RW1900: "double" variable with 6 unique values in the range [0.5, 1.506], 163 NA values.
    \item RW1899: "double" variable with 6 unique values in the range [0.429, 1.048], 163 NA values.
    \item RW1898: "double" variable with 6 unique values in the range [0.463, 1.079], 163 NA values.
    \item RW1897: "double" variable with 6 unique values in the range [0.339, 1.415], 163 NA values.
    \item RW1896: "double" variable with 6 unique values in the range [0.277, 0.674], 163 NA values.
    \item RW1895: "double" variable with 6 unique values in the range [0.192, 0.819], 163 NA values.
    \item RW1894: "double" variable with 6 unique values in the range [0.218, 0.746], 163 NA values.
    \item RW1893: "double" variable with 6 unique values in the range [0.087, 1.123], 163 NA values.
    \item RW1892: "double" variable with 6 unique values in the range [0.247, 2.258], 163 NA values.
    \item RW1891: "double" variable with 6 unique values in the range [0.124, 0.609], 163 NA values.
    \item RW1890: "double" variable with 5 unique values in the range [0.226, 0.555], 164 NA values.
    \item RW1889: "double" variable with 5 unique values in the range [0.203, 0.769], 164 NA values.
    \item RW1888: "double" variable with 5 unique values in the range [0.568, 0.745], 164 NA values.
    \item RW1887: "double" variable with 5 unique values in the range [0.508, 2.074], 164 NA values.
    \item RW1886: "double" variable with 5 unique values in the range [0.087, 0.963], 164 NA values.
    \item RW1885: "double" variable with 3 unique values in the range [0.394, 0.911], 166 NA values.
    \item RW1884: "double" variable with 3 unique values in the range [0.138, 1.134], 166 NA values.
    \item RW1883: "double" variable with 3 unique values in the range [0.314, 0.687], 166 NA values.
    \item RW1882: "double" variable with 3 unique values in the range [0.292, 0.545], 166 NA values.
    \item RW1881: "double" variable with 3 unique values in the range [0.25, 0.814], 166 NA values.
    \item RW1880: "double" variable with 3 unique values in the range [0.172, 0.924], 166 NA values.
    \item RW1879: "double" variable with 3 unique values in the range [0.485, 0.665], 166 NA values.
    \item RW1878: "double" variable with 3 unique values in the range [0.152, 0.643], 166 NA values.
    \item RW1877: "double" variable with 3 unique values in the range [0.153, 0.997], 166 NA values.
    \item RW1876: "double" variable with 3 unique values in the range [0.369, 0.729], 166 NA values.
    \item RW1875: "double" variable with 3 unique values in the range [0.359, 0.621], 166 NA values.
    \item RW1874: "double" variable with 3 unique values in the range [0.341, 0.598], 166 NA values.
    \item YRLASTR: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{1998} \tab \strong{1999}\cr
    116 \tab 51 \tab 1
    }
    \item VAR1: "character" with 9 unique values and 0 NA values.
    \tabular{rrrrrrrrr}{
    \strong{""} \tab \strong{117} \tab \strong{123} \tab \strong{138} \tab \strong{139} \tab \strong{15} \tab \strong{203} \tab \strong{223} \tab \strong{257}\cr
    160 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1
    }
    \item TRCNT: "character" with 9 unique values and 0 NA values.
    \tabular{rrrrrrrrr}{
    \strong{""} \tab \strong{17} \tab \strong{22} \tab \strong{30} \tab \strong{36} \tab \strong{40} \tab \strong{43} \tab \strong{44} \tab \strong{45}\cr
    160 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1
    }
    \item tranid: "character" with 10 unique values and 0 NA values.
    \tabular{rrrrrrrrrr}{
    \strong{""} \tab \strong{E320} \tab \strong{E330} \tab \strong{E334} \tab \strong{E335} \tab \strong{S344} \tab \strong{W32} \tab \strong{W46} \tab \strong{W60} \tab \strong{W70}\cr
    1 \tab 20 \tab 17 \tab 20 \tab 19 \tab 19 \tab 17 \tab 20 \tab 17 \tab 18
    }
    \item plot: "double" variable with 10 unique values in the range [16, 24], 1 NA value.
    \item species: "character" with 2 unique values and 0 NA values.
    \tabular{rr}{
    \strong{""} \tab \strong{BECO}\cr
    1 \tab 167
    }
    \item cii: "double" variable with 8 unique values in the range [1, 5], 22 NA values.
    \item dah1: "double" variable with 128 unique values in the range [0.04, 25.3], 1 NA value.
    \item dah2: "double" variable with 83 unique values in the range [0.04, 8.06], 58 NA values.
    \item dbh1: "double" variable with 62 unique values in the range [0.19, 22], 96 NA values.
    \item dbh2: "double" variable with 18 unique values in the range [0.15, 5.66], 151 NA values.
    \item ht99: "double" variable with 118 unique values in the range [1, 750], 1 NA value.
    \item sub: "character" with 7 unique values and 0 NA values.
    \tabular{rrrrrrr}{
    \strong{""} \tab \strong{1} \tab \strong{3} \tab \strong{CON LITT} \tab \strong{LITTER} \tab \strong{MOSS} \tab \strong{WOOD}\cr
    74 \tab 1 \tab 1 \tab 16 \tab 9 \tab 66 \tab 1
    }
    \item on: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{DEAD WOO} \tab \strong{LIVE WOO} \tab \strong{ROCK} \tab \strong{TIPUP}\cr
    141 \tab 17 \tab 5 \tab 4 \tab 1
    }
    \item term99: "character" with 7 unique values and 0 NA values.
    \tabular{rrrrrrr}{
    \strong{""} \tab \strong{1} \tab \strong{BROWSED} \tab \strong{DEAD} \tab \strong{HEALTHY} \tab \strong{INSECT} \tab \strong{NO DOMIN}\cr
    27 \tab 4 \tab 3 \tab 2 \tab 116 \tab 4 \tab 12
    }
    \item e99: "double" variable with 3 unique values in the range [5.1, 21.4], 166 NA values.
    \item e98: "double" variable with 2 unique values in the range [5.6, 5.6], 167 NA values.
    \item e97: "double" variable with 2 unique values in the range [4.5, 4.5], 167 NA values.
    \item notes: "character" with 19 unique values and 0 NA values.
    \tabular{rrrrrrrrrrrrrrrrrrr}{
    \strong{""} \tab \strong{ANIMAL S} \tab \strong{BEEN BRO} \tab \strong{CORE @ 1} \tab \strong{CORE AND} \tab \strong{CORE AT} \tab \strong{CORE HT} \tab \strong{CORKSCRE} \tab \strong{DAH & CO} \tab \strong{DAH AND} \tab \strong{DAH/CORE} \tab \strong{DEAD TOP} \tab \strong{E98 BROW} \tab \strong{LEANING} \tab \strong{OLD LEAD} \tab \strong{OLDER SH} \tab \strong{ONLY ONE} \tab \strong{TOO SMAL} \tab \strong{VERY TWI}\cr
    144 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1
    }
    \item T1: "double" variable with 11 unique values in the range [250, 1150], 1 NA value.
    \item T2: "double" variable with 11 unique values in the range [250, 1150], 1 NA value.
    \item T3: "double" variable with 11 unique values in the range [250, 1150], 1 NA value.
    \item T4: "double" variable with 10 unique values in the range [250, 1050], 1 NA value.
    \item T5: "double" variable with 11 unique values in the range [250, 1150], 1 NA value.
    \item cht00: "double" variable with 25 unique values in the range [290, 1130], 1 NA value.
}}
\usage{
birchmash_raw
}
\description{
birchmash_raw
}
\keyword{datasets}

