% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birchmasv_raw.R
\docType{data}
\name{birchmasv_raw}
\alias{birchmasv_raw}
\title{birchmasv_raw}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 2708 rows and 30 columns.
\itemize {
    \item DISK: "double" variable with 111 unique values in the range [13, 442], 2329 NA values.
    \item TRAKCORE: "double" variable with 59 unique values in the range [1.1, 80.1], 379 NA values.
    \item YRLASTR: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{1998} \tab \strong{1999}\cr
    116 \tab 2584 \tab 8
    }
    \item YR: "double" variable with 126 unique values in the range [1874, 1998], 116 NA values.
    \item RWIDTH: "double" variable with 1746 unique values in the range [0, 79.456], 116 NA values.
    \item RCNT: "double" variable with 39 unique values in the range [4, 125], 116 NA values.
    \item AGE: "double" variable with 126 unique values in the range [1, 125], 116 NA values.
    \item RYR: "character" with 126 unique values and 0 NA values.
    \item tranid: "character" with 10 unique values and 0 NA values.
    \tabular{rrrrrrrrrr}{
    \strong{""} \tab \strong{E320} \tab \strong{E330} \tab \strong{E334} \tab \strong{E335} \tab \strong{S344} \tab \strong{W32} \tab \strong{W46} \tab \strong{W60} \tab \strong{W70}\cr
    15 \tab 350 \tab 178 \tab 444 \tab 414 \tab 539 \tab 107 \tab 426 \tab 217 \tab 18
    }
    \item plot: "double" variable with 10 unique values in the range [16, 24], 15 NA values.
    \item species: "character" with 2 unique values and 0 NA values.
    \tabular{rr}{
    \strong{""} \tab \strong{BECO}\cr
    15 \tab 2693
    }
    \item cii: "double" variable with 8 unique values in the range [1, 5], 581 NA values.
    \item dah1: "double" variable with 128 unique values in the range [0.04, 25.3], 15 NA values.
    \item dah2: "double" variable with 83 unique values in the range [0.04, 8.06], 2182 NA values.
    \item dbh1: "double" variable with 62 unique values in the range [0.19, 22], 161 NA values.
    \item dbh2: "double" variable with 18 unique values in the range [0.15, 5.66], 2270 NA values.
    \item ht99: "double" variable with 118 unique values in the range [1, 750], 15 NA values.
    \item sub: "character" with 7 unique values and 0 NA values.
    \tabular{rrrrrrr}{
    \strong{""} \tab \strong{1} \tab \strong{3} \tab \strong{CON LITT} \tab \strong{LITTER} \tab \strong{MOSS} \tab \strong{WOOD}\cr
    2481 \tab 1 \tab 1 \tab 16 \tab 48 \tab 160 \tab 1
    }
    \item on: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{DEAD WOO} \tab \strong{LIVE WOO} \tab \strong{ROCK} \tab \strong{TIPUP}\cr
    2660 \tab 17 \tab 5 \tab 25 \tab 1
    }
    \item term99: "character" with 7 unique values and 0 NA values.
    \tabular{rrrrrrr}{
    \strong{""} \tab \strong{1} \tab \strong{BROWSED} \tab \strong{DEAD} \tab \strong{HEALTHY} \tab \strong{INSECT} \tab \strong{NO DOMIN}\cr
    1020 \tab 4 \tab 3 \tab 34 \tab 1280 \tab 99 \tab 268
    }
    \item e99: "double" variable with 3 unique values in the range [5.1, 21.4], 2664 NA values.
    \item e98: "double" variable with 2 unique values in the range [5.6, 5.6], 2665 NA values.
    \item e97: "double" variable with 2 unique values in the range [4.5, 4.5], 2665 NA values.
    \item notes: "character" with 19 unique values and 0 NA values.
    \tabular{rrrrrrrrrrrrrrrrrrr}{
    \strong{""} \tab \strong{ANIMAL S} \tab \strong{BEEN BRO} \tab \strong{CORE @ 1} \tab \strong{CORE AND} \tab \strong{CORE AT} \tab \strong{CORE HT} \tab \strong{CORKSCRE} \tab \strong{DAH & CO} \tab \strong{DAH AND} \tab \strong{DAH/CORE} \tab \strong{DEAD TOP} \tab \strong{E98 BROW} \tab \strong{LEANING} \tab \strong{OLD LEAD} \tab \strong{OLDER SH} \tab \strong{ONLY ONE} \tab \strong{TOO SMAL} \tab \strong{VERY TWI}\cr
    1943 \tab 30 \tab 1 \tab 79 \tab 126 \tab 128 \tab 54 \tab 1 \tab 50 \tab 43 \tab 46 \tab 1 \tab 1 \tab 88 \tab 1 \tab 1 \tab 113 \tab 1 \tab 1
    }
    \item T1: "double" variable with 11 unique values in the range [250, 1150], 15 NA values.
    \item T2: "double" variable with 11 unique values in the range [250, 1150], 15 NA values.
    \item T3: "double" variable with 11 unique values in the range [250, 1150], 15 NA values.
    \item T4: "double" variable with 10 unique values in the range [250, 1050], 15 NA values.
    \item T5: "double" variable with 11 unique values in the range [250, 1150], 15 NA values.
    \item cht00: "double" variable with 25 unique values in the range [290, 1130], 15 NA values.
}}
\usage{
birchmasv_raw
}
\description{
birchmasv_raw
}
\keyword{datasets}

