% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{comp_raw}
\alias{comp_raw}
\title{Archive/compress raw data files}
\usage{
comp_raw(rawfiles, compress_type, basedir = "data-raw", outname = "rawdata",
  program = "7z")
}
\arguments{
\item{rawfiles}{The raw data files to be archived and compressed.}

\item{compress_type}{Compression type {xz, bzip, bzip2, etc}}

\item{basedir}{Name of the raw data directory (default 'data-raw')}

\item{outname}{Basename of output compressed archive (default 'rawdata')}

\item{program}{Path to 7-zip program (7z.exe) on windows (default assumes in path)}
}
\details{
\describe{
   7-zip supports pretty much all the types: http://www.7-zip.org/
   After compression, the raw files/temporary archive are removed
}
}

