% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{grepInOrder}
\alias{grepInOrder}
\title{Grep by year order}
\usage{
grepInOrder(coln, yrs, dat)
}
\arguments{
\item{coln}{Column}

\item{yrs}{Years}

\item{dat}{data}
}
\value{
list of columns
}
\description{
Grep indices ordered by year
Currently this is used in the cleaning of the permanent and transect data.
}

