% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcht15mas_raw.R
\docType{data}
\name{hhcht15mas_raw}
\alias{hhcht15mas_raw}
\title{hhcht15mas_raw}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 194 rows and 10 columns.
\itemize {
    \item TYPE: "character" with 2 unique values and 0 NA values.
    \tabular{rr}{
    \strong{TL10} \tab \strong{TL5}\cr
    83 \tab 111
    }
    \item TRAN: "character" with 9 unique values and 0 NA values.
    \tabular{rrrrrrrrr}{
    \strong{E320} \tab \strong{E330} \tab \strong{E334} \tab \strong{E335} \tab \strong{S344} \tab \strong{W32} \tab \strong{W46} \tab \strong{W60} \tab \strong{W70}\cr
    25 \tab 20 \tab 20 \tab 13 \tab 20 \tab 16 \tab 25 \tab 25 \tab 30
    }
    \item TPLOT: "double" variable with 9 unique values in the range [16, 24], 0 NA values.
    \item ELEV: "double" variable with 38 unique values in the range [1189, 1400], 0 NA values.
    \item AVEELEV: "double" variable with 41 unique values in the range [1224, 1385.495], 0 NA values.
    \item CHT: "double" variable with 105 unique values in the range [1.4, 13.5], 0 NA values.
    \item STDERR: "double" variable with 120 unique values in the range [0, 1.414], 0 NA values.
    \item YEAR: "double" variable with 3 unique values in the range [1999, 2015], 0 NA values.
    \item LOW: "double" variable with 169 unique values in the range [1.324, 13.5], 0 NA values.
    \item UPP: "double" variable with 168 unique values in the range [1.476, 13.5], 0 NA values.
}}
\usage{
hhcht15mas_raw
}
\description{
hhcht15mas_raw
}
\keyword{datasets}

