% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhtr15ht_raw.R
\docType{data}
\name{hhtr15ht_raw}
\alias{hhtr15ht_raw}
\title{hhtr15ht_raw}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 150 rows and 10 columns.
\itemize {
    \item TRAN: "character" with 6 unique values and 0 NA values.
    \tabular{rrrrrr}{
    \strong{E320} \tab \strong{E330} \tab \strong{E335} \tab \strong{W46} \tab \strong{W60} \tab \strong{W70}\cr
    28 \tab 20 \tab 16 \tab 28 \tab 26 \tab 32
    }
    \item TRANSECT: "double" variable with 6 unique values in the range [46, 335], 0 NA values.
    \item TPLOT: "double" variable with 6 unique values in the range [16, 21], 0 NA values.
    \item TAG: "double" variable with 146 unique values in the range [21, 993541], 0 NA values.
    \item SPEC: "character" with 2 unique values and 0 NA values.
    \tabular{rr}{
    \strong{ABBA} \tab \strong{BECO}\cr
    149 \tab 1
    }
    \item HT15: "double" variable with 133 unique values in the range [1.85, 11.1], 0 NA values.
    \item DATE: 
    \item NK: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{JLD}\cr
    150
    }
    \item TECH: "character" with 2 unique values and 0 NA values.
    \tabular{rr}{
    \strong{DRP, PAP, HKS} \tab \strong{HKS}\cr
    122 \tab 28
    }
    \item TCOND15: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{1} \tab \strong{2} \tab \strong{3} \tab \strong{4} \tab \strong{alive and well no dieback}\cr
    32 \tab 90 \tab 24 \tab 3 \tab 1
    }
}}
\usage{
hhtr15ht_raw
}
\description{
hhtr15ht_raw
}
\keyword{datasets}

