% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{levOrder}
\alias{levOrder}
\title{Level ordering}
\usage{
levOrder(v)
}
\arguments{
\item{v}{Variable to convert to factor}
}
\value{
Vector containing order of levels.
}
\description{
If the underlying variable is numeric, like PPLOT,
orders the factor numerically, otherwise uses a standard sort
}

