% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{GPS locations of permanent plots, transect plots, contours and
other random locations of data collected at or near Mt. Moosilauke.}
\format{A data frame (data.table) with 512 and 12 variables.
\itemize{
  \item TPLOT:  Transect plot number (increasing numbers with elevation)
  \item STPACE: Number of paces along the contour.
  \item LABEL:  Notes
  \item lng: Longitude
  \item lat: Latitude
  \item DATE: Date the location data was collected.
  \item POINT_X: Longitude (interpolated or collected)
  \item POINT_Y: Latitude (interpolated or collected)
  \item INTERP: Logical value, whether the location was interpolated or not.
  \item DEMSLOPE: Slope of permanent plot
  \item TYPE: Type of location (permanent plot, transect, etc.)
  \item VALUE: Value depending on type of location (plot number for transect/permanent plot)
}}
\usage{
locations
}
\description{
A dataset containing locations of collected data.  Note that this data
has been separated into separate datasets for permanent plots, transects,
contours, and other locations since they have different pertinent information.
}
\keyword{datasets}

