% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moosedata.R
\name{moosedata}
\alias{moosedata}
\title{Moosilauke data
\itemize{
  \item Remake: Run source('data-raw/remake.R') to remake all the data in 'data' from the
  \item Raw: Raw data in 'data-raw/raw'
}}
\description{
Moosilauke data
\itemize{
  \item Remake: Run source('data-raw/remake.R') to remake all the data in 'data' from the
  \item Raw: Raw data in 'data-raw/raw'
}
}

