% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{other_loc}
\alias{other_loc}
\title{GPS locations of other data locations.}
\format{A data frame (data.table) with 112 rows and 6 variables.
\itemize{
  \item LABEL:  Notes about location
  \item lng: Longitude
  \item lat: Latitude
  \item DATE: Date the location data was collected.
  \item otherLab: HTML Label (used with leaflet for example)
  \item id: Unique identifier of rows (since it is a data.table)
}}
\usage{
other_loc
}
\description{
A dataset containing locations of other (random/unsure) data.
}
\keyword{datasets}

