% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pol2cart}
\alias{pol2cart}
\title{Convert polar coordinates to cartesian}
\usage{
pol2cart(r, theta, deg = FALSE, recycle = FALSE)
}
\arguments{
\item{r}{Radius}

\item{theta}{Angle}

\item{deg}{Degrees (default to FALSE meaning use radians)}

\item{recycle}{Recycle radius values}
}
\value{
cartesian coordinates
}
\description{
Convert polar coordinates to cartesian
}

