% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp}
\alias{pp}
\title{Permanent plot data in long format.}
\format{A data frame with 19011 rows and 43 variables.
\itemize{
  \item PPLOT: Permanent plot number
  \item SPLOT: Subplot number
  \item TAG: Tree tag number
  \item SPEC: Tree species abbreviation
  \item CLASS: ...
}}
\usage{
pp
}
\description{
A dataset containing tree data collected from the 80's to
until most recently in 2011.  The cleaned version in long format.
Some of the variables have been dropped from the raw version.
}
\keyword{datasets}

