% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_demslope.R
\docType{data}
\name{pp_demslope}
\alias{pp_demslope}
\title{Lat/Long locations of Mt. Moosilauke permanent plots and their
slopes guessed from DEM data.}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 27 rows and 4 columns.
\itemize {
    \item PPLOT: "integer" variable with 27 unique values in the range [1, 27], 0 NA values.
    \item LAT: "double" variable with 27 unique values in the range [43.988, 44.015], 0 NA values.
    \item LONG: "double" variable with 26 unique values in the range [-71.858, -71.808], 0 NA values.
    \item SLOPEDEM: "double" variable with 27 unique values in the range [1.024, 34.994], 0 NA values.
}}
\usage{
pp_demslope
}
\description{
Lat/Long locations of Mt. Moosilauke permanent plots and their
slopes guessed from DEM data.
}
\keyword{datasets}

