% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_loc}
\alias{pp_loc}
\title{GPS locations of permananet plot locations.}
\format{A data frame (data.table) with 27 rows and 14 variables.
\itemize{
  \item PPLOT:  Permanent plot number
  \item lng: Longitude
  \item lat: Latitude
  \item DATE: Date the location data was collected.
  \item DEMSLOPE: Slope of permanent plot
  \item ELEVCL: Elevaction class of plot
  \item ASPCL: Aspect class of plot
  \item SOILCL: Soil class of plot
  \item ELEV: Elevation of plot (meters)
  \item ASP: Aspect of plot (degrees)
  \item LABEL:  Additional notes
  \item ppLab: HTML Label (used with leaflet for example)
  \item id: Unique identifier of rows (since it is a data.table)
}}
\usage{
pp_loc
}
\description{
A dataset containing locations of permanent plot data.
}
\keyword{datasets}

