% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_raw}
\alias{pp_raw}
\title{Permanent plot data in raw format (wide).}
\format{A data frame with 12981 rows and 160 variables.
\itemize{
  \item PPLOT: Permanent plot number
  \item SPLOT: Subplot number
  \item TAG: Tree tag number
  \item SPEC: Tree species abbreviation
  \item CLASS: ...
}}
\usage{
pp_raw
}
\description{
A dataset containing tree data collected from the 80's to
until most recently in 2011.
}
\keyword{datasets}

