% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptreemas10_raw.R
\docType{data}
\name{pptreemas10_raw}
\alias{pptreemas10_raw}
\title{pptreemas10_raw}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 12981 rows and 129 columns.
\itemize {
    \item PPLOT: "double" variable with 27 unique values in the range [1, 27], 0 NA values.
    \item SPLOT: "double" variable with 5 unique values in the range [1, 4], 1 NA value.
    \item TAG: "double" variable with 4973 unique values in the range [1, 993983], 0 NA values.
    \item SPEC: "character" with 14 unique values and 0 NA values.
    \tabular{rrrrrrrrrrrrrr}{
    \strong{""} \tab \strong{ABBA} \tab \strong{ACPE} \tab \strong{ACSA} \tab \strong{ACSP} \tab \strong{BEAL} \tab \strong{BECO} \tab \strong{BEPA} \tab \strong{FAGR} \tab \strong{PIRU} \tab \strong{PRPE} \tab \strong{PRVI} \tab \strong{SOAM} \tab \strong{UNID}\cr
    4 \tab 9171 \tab 189 \tab 22 \tab 108 \tab 115 \tab 943 \tab 11 \tab 8 \tab 2105 \tab 31 \tab 2 \tab 255 \tab 17
    }
    \item CLASS: "character" with 4 unique values and 0 NA values.
    \tabular{rrrr}{
    \strong{""} \tab \strong{A} \tab \strong{M} \tab \strong{P}\cr
    1 \tab 1261 \tab 361 \tab 11358
    }
    \item YRTAG: "double" variable with 5 unique values in the range [1986, 2010], 0 NA values.
    \item YRREC: "double" variable with 3 unique values in the range [1998, 2010], 11061 NA values.
    \item YRMORT: "double" variable with 8 unique values in the range [1986, 2011], 9801 NA values.
    \item YRCA: "double" variable with 3 unique values in the range [1988, 2010], 12977 NA values.
    \item CSAP8788: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{N} \tab \strong{Y}\cr
    362 \tab 6626 \tab 5993
    }
    \item CSAP98: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{N} \tab \strong{Y}\cr
    362 \tab 7059 \tab 5560
    }
    \item STAT86: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{ALIVE} \tab \strong{DEAD}\cr
    10586 \tab 1931 \tab 464
    }
    \item STAT87: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{ALIVE} \tab \strong{DEAD}\cr
    10964 \tab 1741 \tab 276
    }
    \item STAT88: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{ALIVE} \tab \strong{DEAD}\cr
    9809 \tab 2668 \tab 504
    }
    \item STAT89: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{ALIVE} \tab \strong{DEAD}\cr
    12653 \tab 317 \tab 11
    }
    \item STAT98: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{ALIVE} \tab \strong{DEAD} \tab \strong{NF} \tab \strong{PD}\cr
    6072 \tab 5285 \tab 1570 \tab 43 \tab 11
    }
    \item STAT10: "character" with 6 unique values and 0 NA values.
    \tabular{rrrrrr}{
    \strong{""} \tab \strong{ALIVE} \tab \strong{DEAD} \tab \strong{NF} \tab \strong{NFC} \tab \strong{PD}\cr
    386 \tab 9424 \tab 2726 \tab 407 \tab 31 \tab 7
    }
    \item STAT11: "character" with 4 unique values and 0 NA values.
    \tabular{rrrr}{
    \strong{""} \tab \strong{ALIVE} \tab \strong{DEAD} \tab \strong{NF}\cr
    12931 \tab 37 \tab 6 \tab 7
    }
    \item DBH86: "double" variable with 253 unique values in the range [5, 50.3], 10585 NA values.
    \item DECM86: "double" variable with 6 unique values in the range [1, 5], 11286 NA values.
    \item HTTCR86: "double" variable with 568 unique values in the range [2.11, 18.7], 11485 NA values.
    \item CRHT86: "double" variable with 555 unique values in the range [0.1, 13.9], 11768 NA values.
    \item CLONG86: "double" variable with 97 unique values in the range [0.6, 13.3], 11966 NA values.
    \item CAZLNG86: "double" variable with 208 unique values in the range [1, 360], 12076 NA values.
    \item CPERP86: "double" variable with 86 unique values in the range [0.3, 10.7], 11966 NA values.
    \item DBH87: "double" variable with 257 unique values in the range [5, 64], 11509 NA values.
    \item HTTCR87: "double" variable with 302 unique values in the range [1, 18.4], 12134 NA values.
    \item TOTHT87: "double" variable with 197 unique values in the range [1, 18.4], 12266 NA values.
    \item CRHT87: "double" variable with 198 unique values in the range [0.33, 10.5], 12699 NA values.
    \item CPOS87: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{c} \tab \strong{d} \tab \strong{i} \tab \strong{o}\cr
    11514 \tab 688 \tab 199 \tab 320 \tab 260
    }
    \item DECM87: "double" variable with 6 unique values in the range [1, 5], 11277 NA values.
    \item DECW87: "double" variable with 6 unique values in the range [1, 5], 11312 NA values.
    \item CLONG87: "double" variable with 79 unique values in the range [1, 14.6], 12758 NA values.
    \item CAZLNG87: "double" variable with 141 unique values in the range [0, 358], 12759 NA values.
    \item CPERP87: "double" variable with 69 unique values in the range [0.8, 12.7], 12758 NA values.
    \item CLITOP87: "double" variable with 76 unique values in the range [3.8, 17.9], 12877 NA values.
    \item HT88: "double" variable with 279 unique values in the range [0.59, 6.12], 11769 NA values.
    \item CPOS88: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{c} \tab \strong{d} \tab \strong{i} \tab \strong{o}\cr
    11072 \tab 735 \tab 186 \tab 773 \tab 215
    }
    \item DECM88: "double" variable with 7 unique values in the range [0, 5], 10704 NA values.
    \item DECW88: "double" variable with 7 unique values in the range [0, 5], 10717 NA values.
    \item WD89OL: "double" variable with 4 unique values in the range [0, 6], 12652 NA values.
    \item WD89AV: "double" variable with 85 unique values in the range [1, 5], 12663 NA values.
    \item EXWI89: "double" variable with 19 unique values in the range [1, 2.583], 12897 NA values.
    \item LVBD89: "double" variable with 36 unique values in the range [0, 12], 12891 NA values.
    \item DECM89: "double" variable with 5 unique values in the range [1, 4], 12653 NA values.
    \item DBH98: "double" variable with 343 unique values in the range [0, 54.5], 7658 NA values.
    \item TOTHT98: "double" variable with 222 unique values in the range [0.4, 21.136], 8068 NA values.
    \item HTBCR98: "double" variable with 123 unique values in the range [0, 11.5], 10412 NA values.
    \item HTTCR98: "double" variable with 222 unique values in the range [0, 21.136], 8068 NA values.
    \item CRHT98: "double" variable with 302 unique values in the range [-5.1, 15.7], 10529 NA values.
    \item DECM98: "double" variable with 5 unique values in the range [1, 4], 9934 NA values.
    \item DECW98: "double" variable with 5 unique values in the range [1, 4], 9932 NA values.
    \item CPOS98: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{c} \tab \strong{d} \tab \strong{i} \tab \strong{o}\cr
    9900 \tab 772 \tab 226 \tab 635 \tab 1448
    }
    \item COND98: "double" variable with 5 unique values in the range [1, 4], 11899 NA values.
    \item SNHT98: "double" variable with 35 unique values in the range [0.05, 10], 12540 NA values.
    \item DECAY98: "double" variable with 7 unique values in the range [1, 6], 11917 NA values.
    \item AOL98: "double" variable with 15 unique values in the range [1, 90], 12901 NA values.
    \item SNAZ98: "double" variable with 137 unique values in the range [5, 360], 12651 NA values.
    \item WD98: "double" variable with 2 unique values in the range [1, 1], 10163 NA values.
    \item ICE98: "double" variable with 2 unique values in the range [1, 1], 9423 NA values.
    \item DATE98: "character" with 92 unique values and 0 NA values.
    \item NOTE98: "character" with 30 unique values and 0 NA values.
    \item DBH10: "double" variable with 918 unique values in the range [0, 58.7], 3591 NA values.
    \item HTTCR10: "double" variable with 855 unique values in the range [0, 23.1], 10648 NA values.
    \item DECM10: "double" variable with 6 unique values in the range [0, 4], 9968 NA values.
    \item CPOS10: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{c} \tab \strong{d} \tab \strong{i} \tab \strong{o}\cr
    9963 \tab 333 \tab 46 \tab 451 \tab 2188
    }
    \item COND10: "double" variable with 6 unique values in the range [1, 7], 11584 NA values.
    \item DECAY10: "double" variable with 10 unique values in the range [1, 11], 11154 NA values.
    \item AOL10: "double" variable with 22 unique values in the range [9, 88], 12953 NA values.
    \item NOTE10: "character" with 16 unique values and 0 NA values.
    \tabular{rrrrrrrrrrrrrrrr}{
    \strong{""} \tab \strong{CA} \tab \strong{CAC} \tab \strong{CS} \tab \strong{NTD} \tab \strong{RL260} \tab \strong{RL265} \tab \strong{RL32} \tab \strong{RL616} \tab \strong{RL654} \tab \strong{RL662} \tab \strong{RL924} \tab \strong{RL925} \tab \strong{RL98818} \tab \strong{RL98823} \tab \strong{top broke of}\cr
    12919 \tab 2 \tab 30 \tab 10 \tab 9 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1
    }
    \item DBH11: "double" variable with 34 unique values in the range [4.4, 39.8], 12947 NA values.
    \item DAH11: "double" variable with 3 unique values in the range [2.6, 3], 12979 NA values.
    \item HTTCR11: "double" variable with 2 unique values in the range [119, 119], 12980 NA values.
    \item CII11: "double" variable with 3 unique values in the range [3, 4], 12979 NA values.
    \item NOTE11: "character" with 4 unique values and 0 NA values.
    \tabular{rrrr}{
    \strong{""} \tab \strong{RL386} \tab \strong{RL522} \tab \strong{RL548}\cr
    12978 \tab 1 \tab 1 \tab 1
    }
    \item ELEV: "double" variable with 23 unique values in the range [808, 1180], 1 NA value.
    \item ELEVCL: "character" with 4 unique values and 0 NA values.
    \tabular{rrrr}{
    \strong{""} \tab \strong{H} \tab \strong{L} \tab \strong{M}\cr
    1 \tab 3081 \tab 2362 \tab 7537
    }
    \item ASPCL: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{E} \tab \strong{W}\cr
    1 \tab 6514 \tab 6466
    }
    \item ASP: "double" variable with 23 unique values in the range [64, 342], 1 NA value.
    \item SOILCL: "character" with 4 unique values and 0 NA values.
    \tabular{rrrr}{
    \strong{""} \tab \strong{.} \tab \strong{H} \tab \strong{S}\cr
    1 \tab 361 \tab 7758 \tab 4861
    }
    \item SLOPE8687: "double" variable with 21 unique values in the range [12, 80], 1 NA value.
    \item SLOPCL: "character" with 5 unique values and 0 NA values.
    \tabular{rrrrr}{
    \strong{""} \tab \strong{BSLP} \tab \strong{FSLP} \tab \strong{SHLD} \tab \strong{SUMT}\cr
    1 \tab 9280 \tab 3339 \tab 165 \tab 196
    }
    \item LAND: "character" with 4 unique values and 0 NA values.
    \tabular{rrrr}{
    \strong{""} \tab \strong{RTOP} \tab \strong{SPUR} \tab \strong{SSLP}\cr
    1 \tab 196 \tab 165 \tab 12619
    }
    \item MICR: "character" with 4 unique values and 0 NA values.
    \tabular{rrrr}{
    \strong{""} \tab \strong{CONC} \tab \strong{CONV} \tab \strong{PLAN}\cr
    1 \tab 3706 \tab 2722 \tab 6552
    }
    \item BQUDX: "double" variable with 13 unique values in the range [0, 11], 1297 NA values.
    \item BQUDY: "double" variable with 13 unique values in the range [0, 11], 1297 NA values.
    \item BQCRDX: "double" variable with 583 unique values in the range [-0.65, 2], 7137 NA values.
    \item BQCRDY: "double" variable with 569 unique values in the range [-1.97, 2.12], 7137 NA values.
    \item BPCRDX: "double" variable with 1612 unique values in the range [-0.65, 22.54], 7169 NA values.
    \item BPCRDY: "double" variable with 1641 unique values in the range [-7.77, 20.77], 7167 NA values.
    \item CLASS98: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{A} \tab \strong{B}\cr
    5723 \tab 6615 \tab 643
    }
    \item EX87: "double" variable with 229 unique values in the range [0.3, 100.2], 12561 NA values.
    \item EX86: "double" variable with 234 unique values in the range [0.3, 95.2], 12571 NA values.
    \item EX85: "double" variable with 229 unique values in the range [0.3, 167], 12576 NA values.
    \item EX84: "double" variable with 227 unique values in the range [0.2, 88.7], 12580 NA values.
    \item EX83: "double" variable with 212 unique values in the range [0.3, 137], 12596 NA values.
    \item OUT: "double" variable with 2 unique values in the range [1, 1], 12937 NA values.
    \item CQUDX86: "double" variable with 13 unique values in the range [0, 11], 11208 NA values.
    \item CQUDY86: "double" variable with 14 unique values in the range [0, 12], 11208 NA values.
    \item CQCRDX86: "double" variable with 156 unique values in the range [0, 2.23], 11226 NA values.
    \item CQCRDY86: "double" variable with 136 unique values in the range [0, 2.12], 11226 NA values.
    \item FLAG: "double" variable with 21 unique values in the range [0, 30], 11308 NA values.
    \item CODE98A: "character" with 36 unique values and 0 NA values.
    \item CENSUS1: "double" variable with 4 unique values in the range [1986, 1988], 7349 NA values.
    \item CENSUS2: "double" variable with 2 unique values in the range [1998, 1998], 5723 NA values.
    \item herb1_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item herb2_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item herb3_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item DECAY: "double" variable with 2 unique values in the range [2, 2], 12980 NA values.
    \item DSAPP88: "double" variable with 7 unique values in the range [0, 6], 8920 NA values.
    \item DSAPA88: "double" variable with 18 unique values in the range [0, 59], 8964 NA values.
    \item DSAPB88: "double" variable with 3 unique values in the range [0, 1], 8920 NA values.
    \item DSAPA87: "double" variable with 10 unique values in the range [0, 11], 11346 NA values.
    \item DSAPP87: "double" variable with 6 unique values in the range [0, 5], 11346 NA values.
    \item DSAPB87: "double" variable with 4 unique values in the range [0, 7], 11346 NA values.
    \item TAGOLD: "double" variable with 1940 unique values in the range [0, 981000], 5730 NA values.
    \item PISAP87: "%s" variable, but all values are NA.
    \item HRB87: "%s" variable, but all values are NA.
    \item HRB1_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item HRB2_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item HRB3_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item SHRB1_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item SHRB2_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item SHRB3_87: "character" with 1 unique values and 0 NA values.
    \tabular{r}{
    \strong{""}\cr
    12981
    }
    \item SPLOT98: "double" variable with 5 unique values in the range [1, 4], 1 NA value.
    \item SPCRT: "double" variable with 2 unique values in the range [1, 1], 12947 NA values.
    \item mdbh8687: "double" variable with 25 unique values in the range [18.6, 50.3], 361 NA values.
    \item mdbh98: "double" variable with 24 unique values in the range [19.7, 54.5], 361 NA values.
    \item mdbh10: "double" variable with 25 unique values in the range [23.2, 58.7], 361 NA values.
}}
\usage{
pptreemas10_raw
}
\description{
pptreemas10_raw
}
\keyword{datasets}

