% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tp_loc}
\alias{tp_loc}
\title{GPS locations of transect plot locations.}
\format{A data frame (data.table) with 237 rows and 13 variables.
\itemize{
  \item TRAN:  Transect name
  \item TPLOT:  Transect plot number
  \item lng: Longitude
  \item lat: Latitude
  \item DATE: Date the location data was collected.
  \item INTERP: Logical value, whether location was interpolated.
  \item ELEVCL: Elevaction class of plot
  \item ASPCL: Aspect class of plot
  \item ELEV: Elevation of plot (meters)
  \item ASP: Aspect of plot (degrees)
  \item LABEL:  Additional notes
  \item tpLab: HTML Label (used with leaflet for example)
  \item id: Unique identifier of rows (since it is a data.table)
}}
\usage{
tp_loc
}
\description{
A dataset containing locations of transect plot data.
}
\keyword{datasets}

