% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tp_raw}
\alias{tp_raw}
\title{Transect plot data in raw format (wide).}
\format{A data frame with 7648 rows and 115 variables.
\itemize{
  \item TRAN: Transect name
  \item TPLOT:  Transect plot number (increasing numbers with elevation)
  \item TAG: Tree tag number
  \item SPEC: Tree species abbreviation
  \item YRTAG: Year the tree was tagged
}}
\usage{
tp_raw
}
\description{
A dataset containing tree data collected from the 80's to
until most recently in 2011.
}
\keyword{datasets}

