% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trspcovmas11_raw.R
\docType{data}
\name{trspcovmas11_raw}
\alias{trspcovmas11_raw}
\title{trspcovmas11_raw}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 2456 rows and 7 columns.
\itemize {
    \item TRAN: "character" with 9 unique values and 0 NA values.
    \tabular{rrrrrrrrr}{
    \strong{E320} \tab \strong{E330} \tab \strong{E334} \tab \strong{E335} \tab \strong{S344} \tab \strong{W32} \tab \strong{W46} \tab \strong{W60} \tab \strong{W70}\cr
    308 \tab 272 \tab 340 \tab 258 \tab 70 \tab 218 \tab 340 \tab 284 \tab 366
    }
    \item TPLOT: "double" variable with 10 unique values in the range [15, 24], 0 NA values.
    \item PCVQ: "character" with 7 unique values and 0 NA values.
    \tabular{rrrrrrr}{
    \strong{""} \tab \strong{       1} \tab \strong{       2} \tab \strong{       3} \tab \strong{L} \tab \strong{R} \tab \strong{U}\cr
    476 \tab 180 \tab 180 \tab 180 \tab 468 \tab 486 \tab 486
    }
    \item SPEC: "character" with 21 unique values and 0 NA values.
    \item PCVR: "double" variable with 36 unique values in the range [0, 95], 0 NA values.
    \item AREA: "double" variable with 1 unique values in the range [1, 1], 0 NA values.
    \item CENS: "double" variable with 2 unique values in the range [1999, 2011], 0 NA values.
}}
\usage{
trspcovmas11_raw
}
\description{
trspcovmas11_raw
}
\keyword{datasets}

