% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/compress.R
\name{comp_raw}
\alias{comp_raw}
\title{Archive/compress raw data files}
\usage{
comp_raw(rawfiles, compress_type, basedir = "data-raw", outname = "rawdata",
  program = "7z")
}
\arguments{
\item{rawfiles}{The raw data files to be archived and compressed.}

\item{compress_type}{Compression type {xz, bzip, bzip2, etc}}

\item{basedir}{Name of the raw data directory (default 'data-raw')}

\item{outname}{Basename of output compressed archive (default 'rawdata')}

\item{program}{Path to 7-zip program (7z.exe) on windows (default assumes in path)}
}
\description{
Compress files with 7-zip if available
}

