% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{decomp_raw}
\alias{decomp_raw}
\title{Decompress/Unarchive rawdata files}
\usage{
decomp_raw(compfile = "rawdata.tar.bzip2", basedir = "data-raw",
  program = "7z")
}
\arguments{
\item{compfile}{Name of the compressed archive or rawdata (default 'rawdata.tar.bzip2')}

\item{basedir}{Name of raw data directory (default 'data-raw')}

\item{program}{Path to 7-zip program for Windows (default assumes on path)}
}
\description{
Decompress files compressed with 7-zip or alternative.
}

