% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{nonEmpty}
\alias{nonEmpty}
\title{Remove null/empty values from list}
\usage{
nonEmpty(lst)
}
\arguments{
\item{lst}{Input list}
}
\value{
List with elements removed
}
\description{
Remove nulls/empty values from list
}

