% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_raw.R
\docType{data}
\name{pp_raw}
\alias{pp_raw}
\title{Permanent plot data in raw format (wide).}
\format{A data frame with 19011 rows and 43 variables.
\itemize{
  \item PPLOT: Permanent plot number
  \item SPLOT: Subplot number
  \item SPLOT98: Subplot number before correction in 2010.
  \item SPCRT: SPCRT=1 if it’s a subplot that got corrected after 1998 basing on BQUDX/BQUDY data.
  \item BQUDX: Quadrat number along X-axis (North -> West).  When missing, same procedure was followed as that for \code{BQCRDX}.
  \item BQUDY: Quadrat number along Y-axis (North -> East).  Same qualifications as \code{BQUDX}.
  \item BQCRDX: Distance along X axis (North -> West) within a 2x2 quadrat from center of the plant.  Only recorded for 1986 tagged trees, 1987 tagged trees and saplings, 1988 tagged saplings, and 1998 tagged trees and saplings.  When missing, and \code{BPCRDX} or \code{BPCRDY} were available, they were calculated from those variables.
  \item BQCRDY: Distance along Y axis (North -> East) within 2x2 quadrat.  Same qualifications as \code{BQCRDX}.
  \item BPCRDX: Distance along X axis (North -> West) from center of the plant within the whole plot.  When missing, this variable was calculated from \code{BPCRDX} and \code{BPCRDY}, or \code{BQUDX} and \code{BQUDY}.
  \item BPCRDY: Distance along Y axis (North -> East) from center of the plant within the whole plot.  Same qualifications as \code{BPCRDX}.
  \item CQUDX86: Quadrat number along X axis (North -> West) from center of tree crown collected in 1986 only.
  \item CQUDY86: Quadrat number along Y axis (North -> East) from center of tree crown collected in 1986 only.
  \item CQCRDX86: Distance along the X axis (North -> West) with 2x2 meter quadrat from the center of the crown.  Collected in 1986 only.
  \item CQCRDY86: Distance along the Y axis (North -> East) with 2x2 meter quadrat from the center of the crown.  Collected in 1986 only.
  \item CSAP8788: Indicates if all saplings in a subplot were tagged ('Y'/'N').
  \item CSAP98: Indicates if all saplings in a subplot were tagged ('Y'/'N').
  \item TAG: Tree tag number used since 1998.  Duplicated tags were marked (add .1 to the duplicated TAG number).
  \item TAGOLD: Old tag numbers used in 1986, 1987, 1988.
  \item SPEC: Tree species abbreviation
  \item CLASS: Type of sampling location
  \itemize{
     \item P: Permanent Plot
     \item A: Additional spruce 'ears'
     \item M: Met station plots.
  }
}}
\usage{
pp_raw
}
\description{
A dataset containing tree data collected from the 80's to
until most recently in 2011.
}
\keyword{datasets}

